move: shadcn/ui 由来のファイルを `ui/ ` 直下に配置
